Allow the package to support PHP 8
This change allows oauth-pkce to support PHP up to version 8.1. As part
of that change, the maximum version of PHPUnit has also been updated to the
most recent release. To accommodate the new version of PHPUnit the tests
have been updated to account for deprecations since the previous maximum
PHPUnit version.